requirements.txt
django-filer
ella>=3.0
ella-galleries
